requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1

# Scheduling
APScheduler==3.10.4

# Configuration Management
PyYAML==6.0.1
python-dotenv==1.0.0

# File Locking (Thread Safety)
filelock==3.13.1

# Data Processing
python-dateutil==2.8.2

# Logging & Monitoring
colorlog==6.8.2

# Optional: For JavaScript-heavy sites (uncomment if needed)
# selenium==4.16.0
# webdriver-manager==4.0.1

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
